+++ src/main/scala/org/mmisw/cf2rdf/cf2rdf.scala
@@ -1,6 +1,6 @@
 package org.mmisw.cf2rdf
 import config.cfg
-import java.io.PrintWriter
+import java.io.{File, PrintWriter}
 
 import org.apache.jena.rdf.model.Model
 import org.apache.jena.system.JenaSystem
@@ -14,9 +14,54 @@ object cf2rdf {
   JenaSystem.init()
 
   def main(args: Array[String]): Unit = {
-    downloadFiles()
-    generateAndSaveRdf()
+    if (args.contains("generate-conf")) {
+      generateConf(args)
+      sys.exit(0)
+    }
+
+    val givenStepNames = collection.mutable.SortedSet[String]()
+    args foreach { stepName ⇒
+      if (!steps.contains(stepName)) {
+        println(s"invalid step $stepName.  Valid steps: ${stepNameOrder.mkString(", ")}")
+        sys.exit(1)
+      }
+      givenStepNames += stepName
+    }
+
+    if (givenStepNames.nonEmpty) {
+      stepNameOrder.filter(givenStepNames.contains) foreach { steps(_)() }
+    }
+    else {
+      println(s"""
+             |Usage:
+             |   cf2rdf generate-conf [--overwrite]
+             |   cf2rdf [download] [convert] [register]
+          """.stripMargin)
+    }
   }
+
+  private def generateConf(args: Array[String]): Unit = {
+    val filename = "cf2rdf.conf"
+    val file = new File(filename)
+    if (file.exists() && !args.contains("--overwrite")) {
+      println(s"$file exists.  Use --overwrite to overwrite")
+      sys.exit(1)
+    }
+    val conf = scala.io.Source.fromInputStream(
+      getClass.getClassLoader.getResource("params_template.conf").openStream()
+    ).mkString
+    val pw = new PrintWriter(file)
+    pw.print(conf)
+    pw.close()
+    println(s" Configuration generated: $filename\n")
+  }
+
+  private val stepNameOrder = List("download", "convert", "register")
+  private val steps = Map[String, () ⇒ Any](
+    "download" → downloadFiles _,
+    "convert"  → generateAndSaveRdf _,
+    "register" → registerOntologies _
+  )
 
   private def downloadFiles(): Unit = {
     download(cfg.xmlUrl, cfg.destXml)
@@ -47,6 +92,16 @@ object cf2rdf {
     println(s"\nSummary: (saved in ${cfg.destStats})\n\t" + statsStr.replaceAll("\n", "\n\t"))
   }
 
+  private def registerOntologies(): Unit = {
+    def requiredEnvVar(name: String): String = {
+      val value = sys.env.getOrElse(name, throw new Exception(s"Undefined $name env var"))
+      value
+    }
+    val userName = requiredEnvVar("ORR_USERNAME")
+    val password = requiredEnvVar("ORR_PASSWORD")
+    cfg.orr foreach { new Registerer(_, userName, password).registerOntologies() }
+  }
+
   private def download(url: String, filename: String): Unit = {
     println(s"Downloading $url")
     val response: HttpResponse[String] = Http(url)
@@ -64,7 +119,7 @@ object cf2rdf {
     val pw = new PrintWriter(createOutputFile(filename))
     pw.print(contents)
     pw.close()
-    println()
+    println(s"            -> $filename\n")
   }
 
   private def saveModel(model: Model) {